copyright + moves conversion function to unit-converter
