+++ src/components/dialogs/network-modifications/line/characteristics-pane/line-characteristics-pane.js
@@ -26,7 +26,7 @@ import {
     SHUNT_SUSCEPTANCE_2,
 } from 'components/utils/field-constants';
 import React, { useEffect, useState } from 'react';
-import { unitToMicroUnit } from 'utils/rounding';
+import { unitToMicroUnit } from 'utils/unit-converter.tsx';
 import { fetchVoltageLevelsListInfos } from '../../../../../services/study/network';
 
 const styles = {
