+++ src/utils/unit-converter.tsx
@@ -1,4 +1,12 @@
-import { kiloUnitToUnit, unitToKiloUnit } from './rounding';
+/**
+ * Copyright (c) 2024, RTE (http://www.rte-france.com)
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+
+import { isBlankOrEmpty } from '../components/utils/validation-functions';
+import { roundToDefaultPrecision } from './rounding';
 
 export enum DISPLAY_CONVERSION {
     NONE, // default mode : data fetched from the back are displayed without any change
@@ -35,3 +43,12 @@ export function convertFromDisplayToRealValues(
             return kiloUnitToUnit(val);
     }
 }
+
+export const unitToMicroUnit = (num: any) =>
+    isBlankOrEmpty(num) ? undefined : roundToDefaultPrecision(num * 1e6);
+export const microUnitToUnit = (num: any) =>
+    isBlankOrEmpty(num) ? undefined : roundToDefaultPrecision(num / 1e6);
+export const kiloUnitToUnit = (num: any) =>
+    isBlankOrEmpty(num) ? undefined : roundToDefaultPrecision(num * 1e3);
+export const unitToKiloUnit = (num: any) =>
+    isBlankOrEmpty(num) ? undefined : roundToDefaultPrecision(num / 1e3);
